+++ firebase-vertexai/consumer-rules.pro
@@ -20,4 +20,5 @@
 # hide the original source file name.
 #-renamesourcefileattribute SourceFile
 
+-keep class com.google.firebase.vertexai.type.** { *; }
 -keep class com.google.firebase.vertexai.common.** { *; }
